# Conflicts:
#	migrations/mysql/SaaS/MigrationContext/MigrationContextModelSnapshot.cs
#	products/ASC.Files/Core/ApiModels/RequestDto/CreateRoomRequestDto.cs
#	products/ASC.Files/Core/Core/Dao/TeamlabDao/FolderDao.cs
#	products/ASC.Files/Core/Core/EF/DbRoomSettings.cs
#	products/ASC.Files/Core/Core/Entries/Folder.cs
#	products/ASC.Files/Core/Core/FileStorageService.cs
#	products/ASC.Files/Core/Core/FilterTypeEnum.cs
#	products/ASC.Files/Core/Core/Security/FileSecurity.cs
#	products/ASC.Files/Core/Core/Thirdparty/ProviderDao/ProviderFolderDao.cs
#	products/ASC.Files/Core/Core/Thirdparty/Sharpbox/SharpBoxFolderDao.cs
#	products/ASC.Files/Core/Core/Thirdparty/ThirdPartyFolderDao.cs
#	products/ASC.Files/Core/Core/Thirdparty/ThirdPartyProviderDao.cs
#	products/ASC.Files/Core/Core/VirtualRooms/RoomLogoManager.cs
#	products/ASC.Files/Core/Helpers/DocSpaceHelper.cs
#	products/ASC.Files/Core/Helpers/DocumentService.cs
#	products/ASC.Files/Core/Helpers/FileUtility.cs
#	products/ASC.Files/Core/Services/DocumentService/DocumentServiceHelper.cs
#	products/ASC.Files/Core/ThirdPartyApp/BoxApp.cs
#	products/ASC.Files/Core/ThirdPartyApp/GoogleDriveApp.cs
#	products/ASC.Files/Server/Api/VirtualRoomsController.cs
#	products/ASC.Files/Service/FileConverterService.cs

+++ common/ASC.Data.Backup.Core/Tasks/BackupPortalTask.cs
@@ -1,4 +1,4 @@
-// (c) Copyright Ascensio System SIA 2010-2023
+// (c) Copyright Ascensio System SIA 2009-2024
 // 
 // This program is a free software product.
 // You can redistribute it and/or modify it under the terms
@@ -110,26 +110,6 @@ public class BackupPortalTask(DbFactory dbFactory,
     {
         var databases = new Dictionary<Tuple<string, string>, List<string>>();
 
-        try
-        {
-            await using var connection = DbFactory.OpenConnection();
-            var command = connection.CreateCommand();
-            command.CommandText = "select id, connection_string from mail_server_server";
-            ExecuteList(command).ForEach(r =>
-            {
-                var connectionString = GetConnectionString((int)r[0], JsonConvert.DeserializeObject<Dictionary<string, object>>(Convert.ToString(r[1]))["DbConnection"].ToString());
-
-                var dbCommand = connection.CreateCommand();
-                dbCommand.CommandText = "show tables";
-                var tables = ExecuteList(dbCommand).Select(objects => Convert.ToString(objects[0])).ToList();
-                databases.Add(new Tuple<string, string>(connectionString.Name, connectionString.ConnectionString), tables);
-            });
-        }
-        catch (Exception e)
-        {
-            logger.ErrorWithException(e);
-        }
-
         await using (var connection = DbFactory.OpenConnection())
         {
             var command = connection.CreateCommand();
@@ -140,7 +120,7 @@ public class BackupPortalTask(DbFactory dbFactory,
 
         using (var stream = new MemoryStream(Encoding.UTF8.GetBytes(true.ToString())))
         {
-            await writer.WriteEntryAsync(KeyHelper.GetDumpKey(), stream, (t) => { });
+            await writer.WriteEntryAsync(KeyHelper.GetDumpKey(), stream, () => Task.CompletedTask);
         }
 
         var files = new List<BackupFileInfo>();
@@ -224,14 +204,15 @@ public class BackupPortalTask(DbFactory dbFactory,
             for (var j = 0; j < TasksLimit && i + j < tables.Count; j++)
             {
                 var t = tables[i + j];
-                tasks.Add(Task.Run(() => DumpTableScheme(t, schemeDir, connectionString)));
-                if (!excluded.Exists(t.StartsWith))
+                var ignore = excluded.Exists(t.StartsWith);
+                tasks.Add(Task.Run(async () => await DumpTableScheme(t, schemeDir, connectionString, ignore)));
+                if (!ignore)
                 {
-                    tasks.Add(Task.Run(() => DumpTableData(t, dataDir, dict[t], connectionString)));
+                    tasks.Add(Task.Run(async () => await DumpTableData(t, dataDir, dict[t], connectionString)));
                 }
                 else
                 {
-                    SetStepCompleted(2);
+                    await SetStepCompleted(2);
                 }
             }
 
@@ -252,7 +233,7 @@ public class BackupPortalTask(DbFactory dbFactory,
 
     }
 
-    private void DumpTableScheme(string t, string dir, string connectionString)
+    private async Task DumpTableScheme(string t, string dir, string connectionString, bool ignore)
     {
         try
         {
@@ -263,21 +244,32 @@ public class BackupPortalTask(DbFactory dbFactory,
                 command.CommandText = $"SHOW CREATE TABLE `{t}`";
                 var createScheme = ExecuteList(command);
                 var creates = new StringBuilder();
-                creates.Append($"DROP TABLE IF EXISTS `{t}`;");
-                creates.AppendLine();
-                creates.Append(createScheme
-                        .Select(r => Convert.ToString(r[1]))
-                        .FirstOrDefault());
-                creates.Append(';');
+                if (ignore)
+                {
+                    creates.Append(createScheme
+                            .Select(r => Convert.ToString(r[1]).Replace("CREATE TABLE ", "CREATE TABLE IF NOT EXISTS "))
+                            .FirstOrDefault());
+                    creates.Append(';');
+                }
+                else
+                {
+
+                    creates.Append($"DROP TABLE IF EXISTS `{t}`;");
+                    creates.AppendLine();
+                    creates.Append(createScheme
+                            .Select(r => Convert.ToString(r[1]))
+                            .FirstOrDefault());
+                    creates.Append(';');
+                }
 
                 var path = CrossPlatform.PathCombine(dir, t);
-                using (var stream = File.OpenWrite(path))
+                await using (var stream = File.OpenWrite(path))
                 {
                     var bytes = Encoding.UTF8.GetBytes(creates.ToString());
                     stream.Write(bytes, 0, bytes.Length);
                 }
 
-                SetStepCompleted();
+                await SetStepCompleted();
             }
 
             logger.DebugDumpTableSchemeStop(t);
@@ -310,14 +302,14 @@ public class BackupPortalTask(DbFactory dbFactory,
 
     }
 
-    private void DumpTableData(string t, string dir, int count, string connectionString)
+    private async Task DumpTableData(string t, string dir, int count, string connectionString)
     {
         try
         {
             if (count == 0)
             {
                 logger.DebugDumpTableDataStop(t);
-                SetStepCompleted(2);
+                await SetStepCompleted(2);
 
                 return;
             }
@@ -400,7 +392,7 @@ public class BackupPortalTask(DbFactory dbFactory,
             } while (true);
 
 
-            SetStepCompleted();
+            await SetStepCompleted();
             logger.DebugDumpTableDataStop(t);
         }
         catch (Exception e)
@@ -457,25 +449,22 @@ public class BackupPortalTask(DbFactory dbFactory,
 
                     for (var i = 0; i < obj.Length; i++)
                     {
-                        var value = obj[i];
-                        if (value is byte[] byteArray && byteArray.Length != 0)
+                        if (obj[i] is byte[] byteArray && byteArray.Length != 0)
                         {
                             sw.Write("0x");
                             foreach (var b in byteArray)
-                            {
                                 sw.Write("{0:x2}", b);
                             }
-                        }
                         else
                         {
                             if (obj[i] is string s)
                             {
-                                sw.Write("'" + s.Replace("\r", "\\r").Replace("\n", "\\n") + "'");
+                                sw.Write("'" + s.Replace("\\", "\\\\").Replace("\r", "\\r").Replace("'", "\\'").Replace("\n", "\\n") + "'");
                             }
                             else
                             {
                                 var ser = new JsonSerializer();
-                                ser.Serialize(writer, value);
+                                ser.Serialize(writer, obj[i]);
                             }
                         }
                         if (i != obj.Length - 1)
@@ -534,7 +523,7 @@ public class BackupPortalTask(DbFactory dbFactory,
         await using (var tmpFile = new FileStream(tmpPath, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read, 4096, FileOptions.DeleteOnClose))
         {
             restoreInfoXml.WriteTo(tmpFile);
-            await writer.WriteEntryAsync(KeyHelper.GetStorageRestoreInfoZipKey(), tmpFile, t => SetStepCompleted());
+            await writer.WriteEntryAsync(KeyHelper.GetStorageRestoreInfoZipKey(), tmpFile, async () => await SetStepCompleted());
         }
         Directory.Delete(subDir, true);
 
@@ -565,7 +554,7 @@ public class BackupPortalTask(DbFactory dbFactory,
             await fileStream.CopyToAsync(tmpFile);
         }
 
-        SetStepCompleted();
+        await SetStepCompleted();
     }
 
     private async Task ArchiveDir(IDataWriteOperator writer, string subDir)
@@ -579,10 +568,8 @@ public class BackupPortalTask(DbFactory dbFactory,
                 f = @"\\?\" + f;
             }
 
-            await using (var tmpFile = new FileStream(f, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read, 4096, FileOptions.DeleteOnClose))
-            {
-                await writer.WriteEntryAsync(enumerateFile[subDir.Length..], tmpFile, task => SetStepCompleted());
-            }
+            await using var tmpFile = new FileStream(f, FileMode.OpenOrCreate, FileAccess.ReadWrite, FileShare.Read, 4096, FileOptions.DeleteOnClose);
+            await writer.WriteEntryAsync(enumerateFile[subDir.Length..], tmpFile, async () => await SetStepCompleted());
         }
 
         logger.DebugArchiveDirEnd(subDir);
@@ -612,55 +599,53 @@ public class BackupPortalTask(DbFactory dbFactory,
             foreach (var table in tablesToProcess)
             {
                 logger.DebugBeginLoadTable(table.Name);
-                using (var data = new DataTable(table.Name))
-                {
-                    ActionInvoker.Try(
-                        state =>
-                        {
-                            data.Clear();
-                            int counts;
-                            var offset = 0;
-                            do
-                            {
-                                var t = (TableInfo)state;
-                                var dataAdapter = DbFactory.CreateDataAdapter();
-                                dataAdapter.SelectCommand = module.CreateSelectCommand(connection.Fix(), TenantId, t, Limit, offset).WithTimeout(600);
-                                counts = ((DbDataAdapter)dataAdapter).Fill(data);
-                                offset += Limit;
-                            } while (counts == Limit);
-
-                        },
-                        table,
-                        maxAttempts: 5,
-                        onFailure: error => throw ThrowHelper.CantBackupTable(table.Name, error),
-                        onAttemptFailure: logger.WarningBackupAttemptFailure);
-
-                    foreach (var col in data.Columns.Cast<DataColumn>().Where(col => col.DataType == typeof(DateTime)))
+                using var data = new DataTable(table.Name);
+                ActionInvoker.Try(
+                    state =>
                     {
-                        col.DateTimeMode = DataSetDateTime.Unspecified;
-                    }
-
-                    module.PrepareData(data);
-
-                    logger.DebugEndLoadTable(table.Name);
-
-                    logger.DebugBeginSavingTable(table.Name);
-
-                    await using (var file = tempStream.Create())
-                    {
-                        data.WriteXml(file, XmlWriteMode.WriteSchema);
                         data.Clear();
+                        int counts;
+                        var offset = 0;
+                        do
+                        {
+                            var t = (TableInfo)state;
+                            var dataAdapter = DbFactory.CreateDataAdapter();
+                            dataAdapter.SelectCommand = module.CreateSelectCommand(connection.Fix(), TenantId, t, Limit, offset).WithTimeout(600);
+                            counts = ((DbDataAdapter)dataAdapter).Fill(data);
+                            offset += Limit;
+                        } while (counts == Limit);
 
-                        await writer.WriteEntryAsync(KeyHelper.GetTableZipKey(module, data.TableName), file, SetProgress);
-                    }
+                    },
+                    table,
+                    maxAttempts: 5,
+                    onFailure: error => throw ThrowHelper.CantBackupTable(table.Name, error),
+                    onAttemptFailure: logger.WarningBackupAttemptFailure);
 
-                    void SetProgress(Task task)
-                    {
-                        SetCurrentStepProgress((int)(++tablesProcessed * 100 / (double)tablesCount));
-                    }
-
-                    logger.DebugEndSavingTable(table.Name);
+                foreach (var col in data.Columns.Cast<DataColumn>().Where(col => col.DataType == typeof(DateTime)))
+                {
+                    col.DateTimeMode = DataSetDateTime.Unspecified;
                 }
+
+                module.PrepareData(data);
+
+                logger.DebugEndLoadTable(table.Name);
+
+                logger.DebugBeginSavingTable(table.Name);
+
+                await using (var file = tempStream.Create())
+                {
+                    data.WriteXml(file, XmlWriteMode.WriteSchema);
+                    data.Clear();
+
+                    await writer.WriteEntryAsync(KeyHelper.GetTableZipKey(module, data.TableName), file, SetProgress);
+                }
+
+                async Task SetProgress()
+                {
+                    await SetCurrentStepProgress((int)(++tablesProcessed * 100 / (double)tablesCount));
+                }
+
+                logger.DebugEndSavingTable(table.Name);
             }
         }
 
@@ -676,9 +661,9 @@ public class BackupPortalTask(DbFactory dbFactory,
             var filesProcessed = 0;
             var filesCount = group.Count();
 
-            void SetProgress(Task task)
+            async Task SetProgress()
             {
-                SetCurrentStepProgress((int)(++filesProcessed * 100 / (double)filesCount));
+                await SetCurrentStepProgress((int)(++filesProcessed * 100 / (double)filesCount));
             }
 
             foreach (var file in group)
@@ -697,7 +682,7 @@ public class BackupPortalTask(DbFactory dbFactory,
         await using (var tmpFile = tempStream.Create())
         {
             restoreInfoXml.WriteTo(tmpFile);
-            await writer.WriteEntryAsync(KeyHelper.GetStorageRestoreInfoZipKey(), tmpFile, task => {});
+            await writer.WriteEntryAsync(KeyHelper.GetStorageRestoreInfoZipKey(), tmpFile, () => Task.CompletedTask);
         }
 
         logger.DebugEndBackupStorage();
